% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_compile.R
\name{clear.criterion.cache}
\alias{clear.criterion.cache}
\title{Clear the live-compiled criterion cache}
\usage{
clear.criterion.cache()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Clear the live-compiled criterion cache
}
\seealso{
Other live-compiled plugins:
\code{\link[=clear.plugin.cache]{clear.plugin.cache()}},
\code{\link[=compile.criterion]{compile.criterion()}},
\code{\link[=compile.criterion.file]{compile.criterion.file()}},
\code{\link[=compile.family]{compile.family()}},
\code{\link[=compile.family.link]{compile.family.link()}},
\code{\link[=compile.link]{compile.link()}},
\code{\link[=list.criterion.cache]{list.criterion.cache()}},
\code{\link[=list.plugin.cache]{list.plugin.cache()}}
}
\concept{live-compiled plugins}
