% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_compile.R
\name{clear.plugin.cache}
\alias{clear.plugin.cache}
\title{Clear the live-compiled plugin cache}
\usage{
clear.plugin.cache()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Clear the live-compiled plugin cache
}
\seealso{
Other live-compiled plugins:
\code{\link[=clear.criterion.cache]{clear.criterion.cache()}},
\code{\link[=compile.criterion]{compile.criterion()}},
\code{\link[=compile.criterion.file]{compile.criterion.file()}},
\code{\link[=compile.family]{compile.family()}},
\code{\link[=compile.family.link]{compile.family.link()}},
\code{\link[=compile.link]{compile.link()}},
\code{\link[=list.criterion.cache]{list.criterion.cache()}},
\code{\link[=list.plugin.cache]{list.plugin.cache()}}
}
\concept{live-compiled plugins}
