% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_compile.R
\name{compile.criterion.file}
\alias{compile.criterion.file}
\title{Live-compile a C++ criterion child from a file}
\usage{
compile.criterion.file(
  file,
  factory,
  name = factory,
  arguments = list(),
  cache = TRUE
)
}
\arguments{
\item{file}{Path to a C++ source file.}

\item{factory}{Name of the \verb{// [[Rcpp::export]]} factory defined by \code{code}.}

\item{name}{Friendly criterion name retained as R metadata.}

\item{arguments}{Named list passed to the exported factory.}

\item{cache}{Reuse an identical compiled criterion from the session cache.}
}
\value{
An external pointer to an \code{IEcountgmifsCriterion} child.
}
\description{
Live-compile a C++ criterion child from a file
}
\seealso{
Other live-compiled plugins:
\code{\link[=clear.criterion.cache]{clear.criterion.cache()}},
\code{\link[=clear.plugin.cache]{clear.plugin.cache()}},
\code{\link[=compile.criterion]{compile.criterion()}},
\code{\link[=compile.family]{compile.family()}},
\code{\link[=compile.family.link]{compile.family.link()}},
\code{\link[=compile.link]{compile.link()}},
\code{\link[=list.criterion.cache]{list.criterion.cache()}},
\code{\link[=list.plugin.cache]{list.plugin.cache()}}
}
\concept{live-compiled plugins}
