% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecountgmifs-package.R
\docType{package}
\name{ecountgmifs-package}
\alias{ecountgmifs-package}
\title{ecountgmifs: Extended count GMIFS}
\description{
Extended generalized monotone incremental forward stagewise regression
for high-dimensional count outcomes.
}
\author{
\strong{Maintainer}: Alikhan Anuarbekov \email{anuarali@fel.cvut.cz}

Authors:
\itemize{
  \item Alikhan Anuarbekov \email{anuarali@fel.cvut.cz}
  \item Jiří Kléma \email{klema@fel.cvut.cz}
}

}
\keyword{internal}
