% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_compile.R
\name{print.ecountgmifs.criterion}
\alias{print.ecountgmifs.criterion}
\title{Print a compiled criterion}
\usage{
\method{print}{ecountgmifs.criterion}(x, ...)
}
\arguments{
\item{x}{object of class \code{"ecountgmifs.criterion"}.}

\item{...}{unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a compiled criterion
}
